Proper format before refactor
The point of this refactor is to simplify the HTTP Task build process. This makes it easier to build more implementations on top later in the development, and it improves the code base's vocabulary.

Currently, a lot of the Elm types have names like 'Snackbar' or 'Context' and they're even threatening to devolve into terms like 'Vnackbar'. Hopefully, many of these types should either be removed or gain renewed names.